Coverity: Fix for CID-1544449, CID-1497428 and CID-1497258 (#11870)
Fix Invalid type in argument to printf format specifier

+++ src/proxy/http/remap/unit-tests/nexthop_test_stubs.cc
@@ -207,5 +207,6 @@ HostStatus::setHostStatus(const std::string_view host, TSHostStatus status, unsi
   this->hosts_statuses[std::string(host)]->status          = status;
   this->hosts_statuses[std::string(host)]->reasons         = reason;
   this->hosts_statuses[std::string(host)]->local_down_time = down_time;
-  NH_Dbg(DbgCtl{"next_hop"}, "setting host status for '%.*s' to %s", host.size(), host.data(), HostStatusNames[status]);
+  NH_Dbg(DbgCtl{"next_hop"}, "setting host status for '%.*s' to %s", static_cast<int>(host.size()), host.data(),
+         HostStatusNames[status]);
 }
